requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
ConfigArgParse==0.15.1
Flask==1.1.1
gevent==1.5a2
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
locustio==0.13.5
MarkupSafe==1.1.1
msgpack-python==0.5.6
psutil==5.6.7
pycparser==2.19
pyzmq==18.1.1
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
